+++ gcc/rust/ast/rust-ast.cc
@@ -35,102 +35,6 @@ along with GCC; see the file COPYING3.  If not see
 namespace Rust {
 namespace AST {
 
-enum indent_mode
-{
-  enter,
-  out,
-  stay
-};
-
-std::string
-indent_spaces (enum indent_mode mode)
-{
-  static int indent = 0;
-  std::string str = "";
-  if (out == mode)
-    indent--;
-  for (int i = 0; i < indent; i++)
-    str += " ";
-  if (enter == mode)
-    indent++;
-
-  return str;
-}
-
-// Gets a string in a certain delim type.
-std::string
-get_string_in_delims (std::string str_input, DelimType delim_type)
-{
-  switch (delim_type)
-    {
-    case PARENS:
-      return "(" + str_input + ")";
-    case SQUARE:
-      return "[" + str_input + "]";
-    case CURLY:
-      return "{" + str_input + "}";
-    default:
-      return "ERROR-MARK-STRING (delims)";
-    }
-  gcc_unreachable ();
-}
-
-enum AttrMode
-{
-  OUTER,
-  INNER
-};
-
-std::string
-get_mode_dump_desc (AttrMode mode)
-{
-  switch (mode)
-    {
-    case OUTER:
-      return "outer attributes";
-    case INNER:
-      return "inner attributes";
-    default:
-      gcc_unreachable ();
-      return "";
-    }
-}
-
-// Adds lines below adding attributes
-std::string
-append_attributes (std::vector<Attribute> attrs, AttrMode mode)
-{
-  indent_spaces (enter);
-
-  std::string str
-    = "\n" + indent_spaces (stay) + get_mode_dump_desc (mode) + ": ";
-  // str += "\n" + indent_spaces (stay) + "inner attributes: ";
-  if (attrs.empty ())
-    {
-      str += "none";
-    }
-  else
-    {
-      /* note that this does not print them with outer or "inner attribute"
-       * syntax - just prints the body */
-      for (const auto &attr : attrs)
-	str += "\n" + indent_spaces (stay) + attr.as_string ();
-    }
-
-  indent_spaces (out);
-
-  return str;
-}
-
-// Removes the beginning and end quotes of a quoted string.
-std::string
-unquote_string (std::string input)
-{
-  rust_assert (input.front () == '"');
-  rust_assert (input.back () == '"');
-  return input.substr (1, input.length () - 2);
-}
-
 std::string
 Crate::as_string () const
 {
@@ -1472,22 +1376,6 @@ ReturnExpr::as_string () const
 }
 
 std::string
-GroupedExpr::as_string () const
-{
-  std::string str ("Grouped expr:");
-
-  // outer attrs
-  str += append_attributes (outer_attrs, OUTER);
-
-  // inner attributes
-  str += append_attributes (inner_attrs, INNER);
-
-  str += "\n Expr in parens: " + expr_in_parens->as_string ();
-
-  return str;
-}
-
-std::string
 RangeToExpr::as_string () const
 {
   return ".." + to->as_string ();
@@ -2461,286 +2349,6 @@ ForLoopExpr::as_string () const
   str += "\n Iterator expr: " + iterator_expr->as_string ();
 
   str += "\n Loop block: " + loop_block->as_string ();
-
-  return str;
-}
-
-std::string
-RangePattern::as_string () const
-{
-  // TODO: maybe rewrite to work with non-linearisable bounds
-  if (has_ellipsis_syntax)
-    return lower->as_string () + "..." + upper->as_string ();
-  else
-    return lower->as_string () + "..=" + upper->as_string ();
-}
-
-std::string
-RangePatternBoundLiteral::as_string () const
-{
-  std::string str;
-
-  if (has_minus)
-    str += "-";
-
-  str += literal.as_string ();
-
-  return str;
-}
-
-std::string
-SlicePattern::as_string () const
-{
-  std::string str ("SlicePattern: ");
-
-  for (const auto &pattern : items)
-    str += "\n " + pattern->as_string ();
-
-  return str;
-}
-
-std::string
-AltPattern::as_string () const
-{
-  std::string str ("AltPattern: ");
-
-  for (const auto &pattern : alts)
-    str += "\n " + pattern->as_string ();
-
-  return str;
-}
-
-std::string
-TuplePatternItemsMultiple::as_string () const
-{
-  std::string str;
-
-  for (const auto &pattern : patterns)
-    str += "\n " + pattern->as_string ();
-
-  return str;
-}
-
-std::string
-TuplePatternItemsRanged::as_string () const
-{
-  std::string str;
-
-  str += "\n Lower patterns: ";
-  if (lower_patterns.empty ())
-    {
-      str += "none";
-    }
-  else
-    {
-      for (const auto &lower : lower_patterns)
-	str += "\n  " + lower->as_string ();
-    }
-
-  str += "\n Upper patterns: ";
-  if (upper_patterns.empty ())
-    {
-      str += "none";
-    }
-  else
-    {
-      for (const auto &upper : upper_patterns)
-	str += "\n  " + upper->as_string ();
-    }
-
-  return str;
-}
-
-std::string
-TuplePattern::as_string () const
-{
-  return "TuplePattern: " + items->as_string ();
-}
-
-std::string
-StructPatternField::as_string () const
-{
-  // outer attributes
-  std::string str = append_attributes (outer_attrs, OUTER);
-
-  return str;
-}
-
-std::string
-StructPatternFieldIdent::as_string () const
-{
-  std::string str = StructPatternField::as_string ();
-
-  str += "\n";
-
-  if (has_ref)
-    str += "ref ";
-
-  if (has_mut)
-    str += "mut ";
-
-  str += ident;
-
-  return str;
-}
-
-std::string
-StructPatternFieldTuplePat::as_string () const
-{
-  // TODO: maybe rewrite to work with non-linearisable patterns
-  std::string str = StructPatternField::as_string ();
-
-  str += "\n";
-
-  str += std::to_string (index) + " : " + tuple_pattern->as_string ();
-
-  return str;
-}
-
-std::string
-StructPatternFieldIdentPat::as_string () const
-{
-  // TODO: maybe rewrite to work with non-linearisable patterns
-  std::string str = StructPatternField::as_string ();
-
-  str += "\n";
-
-  str += ident + " : " + ident_pattern->as_string ();
-
-  return str;
-}
-
-std::string
-StructPatternElements::as_string () const
-{
-  std::string str ("\n  Fields: ");
-
-  if (!has_struct_pattern_fields ())
-    {
-      str += "none";
-    }
-  else
-    {
-      for (const auto &field : fields)
-	str += "\n   " + field->as_string ();
-    }
-
-  str += "\n  Etc: ";
-  if (has_struct_pattern_etc)
-    str += "true";
-  else
-    str += "false";
-
-  return str;
-}
-
-std::string
-StructPattern::as_string () const
-{
-  std::string str ("StructPattern: \n Path: ");
-
-  str += path.as_string ();
-
-  str += "\n Struct pattern elems: ";
-  if (!has_struct_pattern_elems ())
-    str += "none";
-  else
-    str += elems.as_string ();
-
-  return str;
-}
-
-std::string
-LiteralPattern::as_string () const
-{
-  return lit.as_string ();
-}
-
-std::string
-ReferencePattern::as_string () const
-{
-  // TODO: maybe rewrite to work with non-linearisable patterns
-  std::string str ("&");
-
-  if (has_two_amps)
-    str += "&";
-
-  if (is_mut)
-    str += "mut ";
-
-  str += pattern->as_string ();
-
-  return str;
-}
-
-std::string
-IdentifierPattern::as_string () const
-{
-  // TODO: maybe rewrite to work with non-linearisable patterns
-  std::string str;
-
-  if (is_ref)
-    str += "ref ";
-
-  if (is_mut)
-    str += "mut ";
-
-  str += variable_ident;
-
-  if (has_pattern_to_bind ())
-    str += " @ " + to_bind->as_string ();
-
-  return str;
-}
-
-std::string
-TupleStructItemsNoRange::as_string () const
-{
-  std::string str;
-
-  for (const auto &pattern : patterns)
-    str += "\n  " + pattern->as_string ();
-
-  return str;
-}
-
-std::string
-TupleStructItemsRange::as_string () const
-{
-  std::string str ("\n  Lower patterns: ");
-
-  if (lower_patterns.empty ())
-    {
-      str += "none";
-    }
-  else
-    {
-      for (const auto &lower : lower_patterns)
-	str += "\n   " + lower->as_string ();
-    }
-
-  str += "\n  Upper patterns: ";
-  if (upper_patterns.empty ())
-    {
-      str += "none";
-    }
-  else
-    {
-      for (const auto &upper : upper_patterns)
-	str += "\n   " + upper->as_string ();
-    }
-
-  return str;
-}
-
-std::string
-TupleStructPattern::as_string () const
-{
-  std::string str ("TupleStructPattern: \n Path: ");
-
-  str += path.as_string ();
-
-  str += "\n Tuple struct items: " + items->as_string ();
 
   return str;
 }
@@ -4764,12 +4372,6 @@ CompoundAssignmentExpr::accept_vis (ASTVisitor &vis)
 }
 
 void
-GroupedExpr::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
 ArrayElemsValues::accept_vis (ASTVisitor &vis)
 {
   vis.visit (*this);
@@ -5251,138 +4853,6 @@ MacroRulesDefinition::accept_vis (ASTVisitor &vis)
 
 void
 MacroInvocation::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-LiteralPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-IdentifierPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-WildcardPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-RestPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-RangePatternBoundLiteral::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-RangePatternBoundPath::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-RangePatternBoundQualPath::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-RangePattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-ReferencePattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-StructPatternFieldTuplePat::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-StructPatternFieldIdentPat::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-StructPatternFieldIdent::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-StructPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-TupleStructItemsNoRange::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-TupleStructItemsRange::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-TupleStructPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-TuplePatternItemsMultiple::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-TuplePatternItemsRanged::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-TuplePattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-GroupedPattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-SlicePattern::accept_vis (ASTVisitor &vis)
-{
-  vis.visit (*this);
-}
-
-void
-AltPattern::accept_vis (ASTVisitor &vis)
 {
   vis.visit (*this);
 }
